adding missing documentation
