+++ design-library/src/main/java/com/what3words/design/library/ui/components/ResponsiveText.kt
@@ -11,6 +11,23 @@ import androidx.compose.ui.unit.TextUnit
 
 internal const val TEXT_SCALE_REDUCTION_INTERVAL = 0.9f
 
+/**
+ * Composable function to display responsive text that can adapt its size based on the container.
+ *
+ * @param modifier The [Modifier] to be applied to this layout.
+ * @param text The text to be displayed.
+ * @param color The color to apply to the text.
+ * @param textAlign Alignment of the text (default is [TextAlign.Center]).
+ * @param textStyle The style to apply to the text (font, fontWeight, etc.).
+ * @param targetTextSizeHeight The target size of the text, defaults to the font size defined in [textStyle].
+ * @param resizeFunc Optional function to be called when the text needs to resize due to overflowing.
+ * @param maxLines Maximum number of lines the text can occupy, default is 1.
+ *
+ * This function creates a [Text] composable that automatically adjusts its size if the text overflows
+ * and the [resizeFunc] is provided. It's useful for creating text elements that need to respond to
+ * container size changes or dynamic content adjustments. The function respects the maxLines parameter,
+ * truncating the text with an ellipsis if it exceeds the specified line count.
+ */
 @Composable
 internal fun ResponsiveText(
     modifier: Modifier = Modifier,
