+++ design-library/src/main/java/com/what3words/design/library/ui/theme/Color.kt
@@ -14,7 +14,13 @@ import androidx.compose.runtime.staticCompositionLocalOf
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
 
-//primary
+/**
+ * Defines the custom color palette for What3words, extending the MaterialTheme color scheme.
+ * This includes both light and dark theme colors and additional semantic colors for specific UI components.
+ * These colors can be overridden in specific parts of the app using local composition.
+ */
+
+// Primary color definitions for light and dark themes.
 val w3w_theme_light_primary = Color(0xFFC00004)
 val w3w_theme_light_onPrimary = Color(0xFFFFFFFF)
 val w3w_theme_light_primaryContainer = Color(0xFFFFCAC2)
@@ -25,7 +31,7 @@ val w3w_theme_dark_onPrimary = Color(0xFF690001)
 val w3w_theme_dark_primaryContainer = Color(0xFF930002)
 val w3w_theme_dark_onPrimaryContainer = Color(0xFFFFCAC2)
 
-//secondary
+// Secondary color definitions for light and dark themes.
 val w3w_theme_light_secondary = Color(0xFF006397)
 val w3w_theme_light_onSecondary = Color(0xFFFFFFFF)
 val w3w_theme_light_secondaryContainer = Color(0xFFDBEFFA)
@@ -36,7 +42,7 @@ val w3w_theme_dark_onSecondary = Color(0xFF0A3049)
 val w3w_theme_dark_secondaryContainer = Color(0xFF005379)
 val w3w_theme_dark_onSecondaryContainer = Color(0xFFDBEFFA)
 
-//error
+// Error color definitions for light and dark themes.
 val w3w_theme_light_error = Color(0xFFF26C50)
 val w3w_theme_light_errorContainer = Color(0xFFFFDAD3)
 val w3w_theme_light_onError = Color(0xFFFFFFFF)
@@ -47,7 +53,7 @@ val w3w_theme_dark_errorContainer = Color(0xFFCA4F36)
 val w3w_theme_dark_onError = Color(0xFF640D00)
 val w3w_theme_dark_onErrorContainer = Color(0xFFFFDAD3)
 
-//warning (what3words custom)
+// Warning color definitions for light and dark themes, specific to What3words.
 val w3w_theme_light_warning = Color(0xFFC7AA00)
 val w3w_theme_light_warningContainer = Color(0xFFFFF4B2)
 val w3w_theme_light_onWarning = Color(0xFFFFFFFF)
@@ -58,7 +64,7 @@ val w3w_theme_dark_warningContainer = Color(0xFFC7AA00)
 val w3w_theme_dark_onWarning = Color(0xFF6F5D00)
 val w3w_theme_dark_onWarningContainer = Color(0xFFFFF4B2)
 
-//success (what3words custom)
+// Success color definitions for light and dark themes, specific to What3words.
 val w3w_theme_light_success = Color(0xFF008857)
 val w3w_theme_light_successContainer = Color(0xFF8BF8BD)
 val w3w_theme_light_onSuccess = Color(0xFFFFFFFF)
@@ -69,7 +75,7 @@ val w3w_theme_dark_successContainer = Color(0xFF006C45)
 val w3w_theme_dark_onSuccess = Color(0xFF003822)
 val w3w_theme_dark_onSuccessContainer = Color(0xFF8BF8BD)
 
-//surface
+// Surface color definitions for light and dark themes.
 val w3w_theme_light_surface = Color(0xFFFBFDF9)
 val w3w_theme_light_onSurface = Color(0xFF191C1A)
 val w3w_theme_light_surfaceVariant = Color(0xFFEBEBEB) //missing map
@@ -95,14 +101,14 @@ val w3w_theme_dark_surfaceContainerLow = Color(0xFF1A1C1E)
 val w3w_theme_dark_surfaceContainerHighest = Color(0xFF333537)
 val w3w_theme_dark_surfaceContainerLowest = Color(0xFF0C0E11)
 
-//background
+// Background color definitions for light and dark themes.
 val w3w_theme_light_background = Color(0xFFFFFBFF)
 val w3w_theme_light_onBackground = Color(0xFF1A1C1E)
 
 val w3w_theme_dark_background = Color(0xFF1A1C1E)
 val w3w_theme_dark_onBackground = Color(0xFFE2E2E5)
 
-//misc
+// Miscellaneous color definitions for light and dark themes.
 val w3w_theme_light_outline = Color(0xFFEBEBEB)
 val w3w_theme_light_inverseOnSurface = Color(0xFFF0F0F3)
 val w3w_theme_light_inverseSurface = Color(0xFF2E3133)
@@ -119,6 +125,9 @@ val w3w_theme_dark_shadow = Color(0xFF000000) //not in lib
 val w3w_theme_dark_outlineVariant = Color(0xFF32373C)
 val w3w_theme_dark_scrim = Color(0xFF000000)
 
+/**
+ * Defines a custom color scheme for What3words using the defined colors for light themes.
+ */
 internal val w3wLightColors = lightColorScheme(
     primary = w3w_theme_light_primary,
     onPrimary = w3w_theme_light_onPrimary,
@@ -153,6 +162,9 @@ internal val w3wLightColors = lightColorScheme(
     surfaceContainerLowest = w3w_theme_light_surfaceContainerLowest
 )
 
+/**
+ * Defines a custom color scheme for What3words using the defined colors for dark themes.
+ */
 internal val w3wDarkColors = darkColorScheme(
     primary = w3w_theme_dark_primary,
     onPrimary = w3w_theme_dark_onPrimary,
@@ -187,6 +199,9 @@ internal val w3wDarkColors = darkColorScheme(
     surfaceContainerLowest = w3w_theme_dark_surfaceContainerLowest
 )
 
+/**
+ * Data class representing custom success colors, allowing for additional color customization.
+ */
 @Immutable
 data class SuccessColors(
     val success: Color = Color.Unspecified,
@@ -195,6 +210,7 @@ data class SuccessColors(
     val onSuccessContainer: Color = Color.Unspecified
 )
 
+// Success color assignments for light and dark themes.
 val lightSuccessColors = SuccessColors(
     success = w3w_theme_light_success,
     onSuccess = w3w_theme_light_onSuccess,
@@ -209,9 +225,9 @@ val darkSuccessColors = SuccessColors(
     onSuccessContainer = w3w_theme_dark_onSuccessContainer
 )
 
-val LocalDarkSuccessColors = staticCompositionLocalOf { darkSuccessColors }
-val LocalLightSuccessColors = staticCompositionLocalOf { lightSuccessColors }
-
+/**
+ * Data class representing custom warning colors, allowing for additional color customization.
+ */
 @Immutable
 data class WarningColors(
     val warning: Color = Color.Unspecified,
@@ -220,6 +236,7 @@ data class WarningColors(
     val onWarningContainer: Color = Color.Unspecified
 )
 
+// Warning color assignments for light and dark themes.
 val lightWarningColors = WarningColors(
     warning = w3w_theme_light_warning,
     onWarning = w3w_theme_light_onWarning,
@@ -234,27 +251,18 @@ val darkWarningColors = WarningColors(
     onWarningContainer = w3w_theme_dark_onWarningContainer
 )
 
+/**
+ * Composition locals for custom success and warning colors.
+ */
 val LocalDarkWarningColors = staticCompositionLocalOf { darkWarningColors }
 val LocalLightWarningColors = staticCompositionLocalOf { lightWarningColors }
+val LocalDarkSuccessColors = staticCompositionLocalOf { darkSuccessColors }
+val LocalLightSuccessColors = staticCompositionLocalOf { lightSuccessColors }
 
-@Immutable
-data class W3wTypography(
-    val titleMediumProminent: TextStyle,
-    val labelLargeProminent: TextStyle,
-    val labelMediumProminent: TextStyle
-)
-
-val LocalW3wTypography = staticCompositionLocalOf<W3wTypography?> { null }
-
-val MaterialTheme.w3wTypography: W3wTypography
-    @Composable
-    @ReadOnlyComposable
-    get() = LocalW3wTypography.current ?: W3wTypography(
-        titleMediumProminent = MaterialTheme.typography.titleMedium,
-        labelLargeProminent = MaterialTheme.typography.labelLarge,
-        labelMediumProminent = MaterialTheme.typography.labelMedium
-    )
-
+/**
+ * Extension properties on [MaterialTheme] to provide easy access to custom warning and success colors.
+ * The theme automatically switches between light and dark color sets based on the system theme.
+ */
 val MaterialTheme.warningColors: WarningColors
     @Composable
     @ReadOnlyComposable
